pq3: add Sync, support dual FE/BE message types
Sync and ParameterStatus share a type byte. This means that we need to
know whether we're a client or a server as a context clue to be able to
correctly decode packets.

Add support for this in the EnumNamedByte/ByteEnum infrastructure, and
mark a wrapped pq3 stream as a client automatically if send_startup() is
called on it.